Improve collection listing (#70)
Collection pages now have a much more useful listing, they display a
heading for each page in the collection, a description if present, and a
call to action which links through to the page. This has definitely
highlighted that most of our pages have no description.

There is a new frontmatter property which suppresses this behaviour on a
page-by-page basis, allowing us to hide items from the listing where
they don't make sense (e.g. the markup/js/scss specs). The frontmatter
is named `collection_listing_display` and it defaults to `true`.

This closes #64