Write meta descriptions and CTAs for all pages
This introduces meta descriptions for all pages. This is useful for
search-engine optimisation, but also we use these descriptions wherever
the page is listed on the website.

This also introduces a call-to-action for all collection pages. The CTA
is used in page listings instead of generic link text.

This closes #61