Add redirects from all the old site URLs (#76)
This ensures that bookmarks etc that people may have will continue to
work. There are a few less-than-ideal redirects, e.g. using the polyfill
service with components used to be a full page but it now redirects to a
page that has a small sub-section on this topic, however there's no way
to redirect to that specific page section in Jekyll. I think this is
better than no redirect though.

This closes #75.

There are a few pages on the old site which aren't a simple redirect. I
think these should be discussed separately and addressed as separate
issues (which I'll be opening):

  * [syntax overview](https://origami.ft.com/docs/syntax/)
  * [web service description format](https://origami.ft.com/docs/syntax/web-service-description/)
  * [web service index format](https://origami.ft.com/docs/syntax/web-service-index/)
  * [metrics page](https://origami.ft.com/docs/syntax/metrics/)
  * [third party components](https://origami.ft.com/docs/3rd-party-a-list/)
  * [super quick start](https://origami.ft.com/docs/developer-guide/modules/very-quick-origami/)