Add a blog post outlining survey results
I've added a blog post outlining recent survey results. I've avoided
drawing too many conclusions, and think we can update when the whole
team's had a chance to look over the results and we've discussed them
in a planning session.

The charts are a little bit scappy, I've tried to make them a bit
generic and reusable but realistically this is going to need more work.
I suggest we try to use them next time we publish survey results and
make some usability changes then.