Add defer attribute to blocking script (#199)
Small change to improve the performance of the tutorial page. Typically best practice is to put scripts at the bottom of the page, this is because they will block the parsing of HTML slowing down the initial render of the page. Given that this is a tutorial it makes sense to keep these in the `<head>` to keep things simple, we can then use the `defer` attribute to avoid blocking render in newer browsers.