Update component spec url structure and breadcrumbs
Move sass, js, and markup spec under the component spec in the
url structure:
- `/spec/v1/sass/` to `/spec/v1/components/sass/`
- `/spec/v1/javascript/` to `/spec/v1/components/javascript/`
- `/spec/v1/markup/` to `/spec/v1/components/markup/`

And update the breadcrumbs in the navigation accordingly.

_This is messy code which follows our existing approach. I don't know if
there is a better way with jekyll. We may want to move away from jekyll
to something more familiar to the team in the future._

Addresses: https://github.com/Financial-Times/origami-website/issues/187