Move demos specification into its own file
This moves the description of how component demos are defined into their own
specification, transferring content from the manifest and components
specification.

The intention is to allow us to continue pointing to this description of how
demos work in v2.0, and then when we do a period of discovery on future demos we
can do a minor spec v2 release that adds support for the new demo spec (while
keeping support for the old demo type). This way we avoid creating a series of
major releases for our consumers when we change how demos work in the future.

I've written more about this here:
https://github.com/Financial-Times/origami/pull/86#discussion_r513582187

Unfortunately, this does introduce a bit of indirection in the manifest
specification but I think that is OK.

Current thoughts are that once we have a specification for demos v2, the value
of the "demos" property in an origami.json will be "v2".

{
  "demos": "v2"
}