Remove `_specification-v2/sass.md`
The v2 spec is being worked on in the v1 directory to display a
diff in Github. When it is complete it will be moved into the
`_specification-v2` directory, and v1 of the spec restored in
`_specification-v1`.

Somehow `_specification-v2/sass.md` was left behind and had become
out of date.