Require V2 components are published under a Financial Times npm scope.
Currently that is only `@financial-times` but may include more npm scopes
in the future. This is so tools and services may be more confident in the
contents of components, from a security point of view. If The Build
Service was required to compile any npm package a malicious component could
be released by an attacker with demos and assets hosted under the `ft.com`
domain name. A limited set of allowed scopes, owned by the FT, will
mitigate such a scenario without increasing the complexity of our
tools/services.