Wrap code examples in email (#328)
Currently code blocks may overflow and break email templates. This
commit makes it so code wraps rather than overflows the email
container. I went with this over scrolling overflow as we have no
code block styles for email sends and it's not obvious the blocks
can be scrolled. If we ever need to share large code blocks via
email we may want to update the styles and make overflow scroll,
but it's more likely right now we're sharing inline code blocks
or something simple like a single command or url.