Note that the Origami specification is deprecated.
This is a bit grim, as we don't have the alternative yet.
We decided to deprecate the spec now to avoid a second
cascade of component releases in the near future, which
can be a challenge to roll out across Customer Products.

TODO: add links to a blog post with an explanation.