Add non-normative component "code" documentation.
For reference whilst we have deprecated the spec but not figured
out what's next entirely.

Takes the latest spec v2 draft but makes it non-normative
and noncommittal. I tried starting from spec v1, to retain some
content we decided was guidence that does not belong in a spec,
however found most of it not very useful to have here even --
more tutorial and workshop content.
https://github.com/Financial-Times/origami-website/pull/359